feat: use typescript and changd the flow to set config values
