+++ package.json
@@ -2,7 +2,8 @@
   "name": "razorpayx-nodejs-sdk",
   "version": "1.2.0",
   "description": "This is implementation of razorpayx api in NodeJS",
-  "main": "src/index.js",
+  "main": "./dist/index.js",
+  "types": "./dist/index.d.ts",
   "scripts": {
     "build": "tsc",
     "release": "semantic-release"
