--- src/index.js
@@ -1,9 +1,0 @@
-const RazorPayX = {
-  RazorPayContact: require("./Contact"),
-  RazorPayFundAccount: require("./FundAccount"),
-  RazorPayPayout: require("./Payout"),
-  RazorPayPayoutLink: require("./PayoutLink"),
-  RazorPayTransactions: require("./Transactions"),
-};
-
-module.exports = RazorPayX;
